+++ packages/core/src/lowcode-validation/index.ts
@@ -35,6 +35,12 @@ export {
 
 export { decodeJwtPayload, detectServiceRole, validateSupabaseConfig } from './supabase-config'
 
+export {
+  lowcodeCustomCssUrls,
+  unsafeLowcodeCustomCssUrls,
+  validateLowcodeCustomCss
+} from './custom-css'
+
 export { PAYLOAD_ENTRY_KEY_RE, validateSupabasePayloadEntries } from './supabase-payload-entries'
 
 export {